Split SqlCommandReader taking the parsing bits out into a new SqlParser
